Support decimal NDS tile heights

+++ openmmo-map-editor/src/main/kotlin/de/lananahwp/openmmo/mapeditor/ui/Nds3DView.kt
@@ -254,7 +254,7 @@ interface Nds3DView {
   var modelOpacity: Float
   var activeLayer: Int
   var activeTile: Int
-  var activeHeight: Int
+  var activeHeight: Double
   var brushCollision: Int
   var showGrid: Boolean
   var showCollision: Boolean
